Style stop watch and stop watch buttons.
