+++ package.json
@@ -13,7 +13,8 @@
   "license": "ISC",
   "dependencies": {
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "sass": "^1.70.0"
   },
   "devDependencies": {
     "@babel/core": "^7.21.8",
@@ -27,13 +28,14 @@
     "@types/react-dom": "^18.2.4",
     "babel-jest": "^29.7.0",
     "babel-loader": "^8.3.0",
-    "css-loader": "^6.7.3",
+    "css-loader": "^6.9.1",
     "file-loader": "^6.2.0",
     "html-loader": "^4.2.0",
     "html-webpack-plugin": "^5.5.1",
     "jest": "^29.7.0",
     "react-test-renderer": "^18.2.0",
-    "style-loader": "^3.3.2",
+    "sass-loader": "^14.0.0",
+    "style-loader": "^3.3.4",
     "ts-jest": "^29.1.1",
     "ts-loader": "^9.4.2",
     "typescript": "^4.9.5",
