+++ src/components/StopWatchButton/StopWatchButton.scss
@@ -1,0 +1,20 @@
+.button {
+
+    border-radius: 20px;
+    width: 12rem;
+    height: 2.5rem;
+    margin: 0 1rem;
+    border: none;
+    background-color: #50c0e5;
+    cursor: pointer;
+    font-size: 1.5rem;
+
+    &:hover {
+        background-color: #316f83;
+    }
+
+    &__container {
+        display: flex;
+        justify-content: space-evenly;
+    }
+}
