+++ webpack.common.js
@@ -18,32 +18,35 @@ module.exports = {
   },
   module: {
     rules: [{
-        test: /\.tsx?$/,
-        use: 'ts-loader',
-        exclude: /node_modules/,
-    },{
-        test: /\.jsx?$/,
-        exclude: /node_modules/,
-        loader: 'babel-loader',
-        options: {
-            presets: ['@babel/preset-env', '@babel/preset-react', '@babel/plugin-syntax-jsx', '@babel/preset-typescript'],
-        }
-    },{
-        test: /\.css$/,
-        use: [ 'style-loader', 'css-loader' ]
-    },{
+      test: /\.tsx?$/,
+      use: 'ts-loader',
+      exclude: /node_modules/,
+    }, {
+      test: /\.jsx?$/,
+      exclude: /node_modules/,
+      loader: 'babel-loader',
+      options: {
+        presets: ['@babel/preset-env', '@babel/preset-react', '@babel/plugin-syntax-jsx', '@babel/preset-typescript'],
+      }
+    }, {
+      test: /\.css$/,
+      use: ['style-loader', 'css-loader']
+    }, {
       test: /\.(png|jpe?g|gif|jp2|webp)$/,
       loader: 'file-loader',
       options: {
         name: 'images/[name].[ext]'
       }
-    }
-  ],
-},
-resolve: {
-  extensions: ['.tsx', '.ts', '.js'],
-},
-devServer: {
-  historyApiFallback: true,
-},
+    }, {
+      test: /\.scss$/,
+      use: ['style-loader', 'css-loader', 'sass-loader'],
+    },
+    ],
+  },
+  resolve: {
+    extensions: ['.tsx', '.ts', '.js'],
+  },
+  devServer: {
+    historyApiFallback: true,
+  },
 };
